Basic Playermovement
I have script for a solid cube that has rigidbody2D, and can be controlled via WASD/Arrow Keys to travel along the XY Plane. Note that unity assumes Y is the Up/Down axis so we could make our game along this plane or restructure the code to move the player along XZ but I will have to look into the terminology for this.